optimizations

+++ src/main/java/com/nexia/core/mixin/entity/LivingEntityMixin.java
@@ -68,7 +68,7 @@ public abstract class LivingEntityMixin {
             int ping = 0;
             if (instance instanceof ServerPlayer)
                 ping = ((ServerPlayer)instance).latency;
-            double p = ping/50;
+            double p = (double) ping /50;
             double m = p * -0.0784/vec3.y;
             double mi = 1-m;
             instance.setDeltaMovement(vec3.x / 2.0 - vec32.x, instance.isOnGround() ? Mth.clamp((double)f * 0.75, 0.0, 0.4) : Mth.clamp(2/3 * vec3.y + (double)f * 0.5, -0.4, 0.4), vec3.z / 2.0 - vec32.z);
